Add property concept occurrence model.
Instances of this type consist of a concept from the controlled vocabulary along with
an external property and an external concept. This represents that during the classification
of a named entity, that external property has been found on the corresponding wikidata item page
in combination with an instance of the external concept.
The cumulative list of property concept occurrences can be queried for recommendation generation.