Prepare more generic way of additional property retrieval.
Those properties and their internal aliases that we want to retrieve in order to show their values in the
annotator search result menu are now being defined in a dictionary which could easily be exposed as a configuration
in the app's settings if it seems reasonable.

Also, make final search result item schema mapping delete the 'match' part of wikibase API responses in order to limit
local http traffic.